requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
python-twitch-client>=0.7.1
aiohttp>=3.8.1